Move layouts under application
These layouts are specific to this application and as such not
necessarily common resources.
