+++ src/domain/application/layout/kukkuuAppLayout/KukkuuAppLayout.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Layout } from 'react-admin';
 
-import KukkuuAppBar from '../appBar/KukkuuAppBar';
+import KukkuuAppBar from '../../../../common/components/appBar/KukkuuAppBar';
 
 const KukkuuLayout = (props: any) => (
   <Layout {...props} appBar={KukkuuAppBar} />
