+++ src/domain/application/layout/kukkuuCardPageLayout/KukkuuCardPageLayout.tsx
