+++ src/domain/application/layout/kukkuuDetailPage/KukkuuDetailPage.tsx
