+++ src/domain/application/layout/kukkuuEdit/KukkuuEdit.tsx
@@ -1,7 +1,7 @@
 import { Edit } from 'react-admin';
 import { styled } from '@material-ui/core/styles';
 
-import { kukkuuToolbar } from '../../materialUI/kukkuuStyles';
+import { kukkuuToolbar } from '../../../../common/materialUI/kukkuuStyles';
 
 const KukkuuEdit = styled(Edit)(kukkuuToolbar);
