+++ src/domain/application/layout/kukkuuEditPage/KukkuuEditPage.tsx
@@ -2,7 +2,7 @@ import React, { ReactElement } from 'react';
 import { ResourceComponentPropsWithId } from 'react-admin';
 
 import KukkuuPageLayout from '../kukkuuCardPageLayout/KukkuuCardPageLayout';
-import KukkuuEdit from '../../../common/components/kukkuuEdit/KukkuuEdit';
+import KukkuuEdit from '../kukkuuEdit/KukkuuEdit';
 
 type Props = {
   reactAdminProps: ResourceComponentPropsWithId;
