+++ src/domain/application/layout/kukkuuList/KukkuuList.tsx
@@ -1,7 +1,7 @@
 import { List } from 'react-admin';
 import { styled } from '@material-ui/core/styles';
 
-import { kukkuuToolbar } from '../../materialUI/kukkuuStyles';
+import { kukkuuToolbar } from '../../../../common/materialUI/kukkuuStyles';
 
 const KukkuuList = styled(List)(kukkuuToolbar);
