+++ src/domain/application/layout/kukkuuShow/KukkuuShow.tsx
@@ -1,7 +1,7 @@
 import { Show } from 'react-admin';
 import { styled } from '@material-ui/core/styles';
 
-import { kukkuuToolbar } from '../../materialUI/kukkuuStyles';
+import { kukkuuToolbar } from '../../../../common/materialUI/kukkuuStyles';
 
 const KukkuuShow = styled(Show)(kukkuuToolbar);
