+++ src/domain/messages/edit/MessagesEdit.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ResourceComponentProps } from 'react-admin';
 
-import KukkuuEditPage from '../../../common/components/kukkuuEditPage/KukkuuEditPage';
+import KukkuuEditPage from '../../application/layout/kukkuuEditPage/KukkuuEditPage';
 import MessageForm from '../form/MessageForm';
 
 const MessagesEdit = (props: ResourceComponentProps) => {
